// src/modules/33-auth-settings/pages/subscriptions/overview/types.ts
/*
 * Copyright 2021 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Shield 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
 */

export interface SortBy {
  sort: 'serviceInstances' | 'lastDeployed' | 'licensesConsumed'
  order: 'DESC' | 'ASC'
}